좌표 float에서 double로 변경

+++ Assets/02.Scripts/AR 길안내 페이지 스크립트/ARDirectionsManager.cs
@@ -25,8 +25,8 @@ public class ARDirectionsManager : GPS
         Request();
     }
 
-    float myLat = 0f;
-    float myLong = 0f;
+    double myLat = 0f;
+    double myLong = 0f;
     void Update()
     {
         test();
